scrapers/Linkedin_API/requirements.txt
# Core Web Framework
Flask==2.3.3
flask-restx==1.1.0
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
werkzeug==2.3.7

# Web Automation (LinkedIn Scraping)
selenium==4.12.0
undetected-chromedriver==3.5.3
webdriver-manager==4.0.1

# Database
pymysql==1.1.0
cryptography==41.0.7

# Environment & Config
python-dotenv==1.0.0

# Utilities
pyperclip==1.8.2
requests==2.31.0
python-dateutil==2.8.2

# Telegram Notifications (Optional)
python-telegram-bot==20.6

# Production Server
gunicorn==21.2.0
